First pass of client credentials flow
