+++ src/fidesops/ops/api/v1/endpoints/oauth_endpoints.py
@@ -45,8 +45,8 @@ from fidesops.ops.schemas.client import ClientCreatedResponse
 from fidesops.ops.service.authentication.authentication_strategy_factory import (
     get_strategy,
 )
-from fidesops.ops.service.authentication.authentication_strategy_oauth2 import (
-    OAuth2AuthenticationStrategy,
+from fidesops.ops.service.authentication.authentication_strategy_oauth2_auth_code import (
+    OAuth2AuthCodeAuthenticationStrategy,
 )
 from fidesops.ops.util.api_router import APIRouter
 from fidesops.ops.util.oauth_util import verify_oauth_client
@@ -214,9 +214,10 @@ def oauth_callback(code: str, state: str, db: Session = Depends(get_db)) -> None
         authentication = (
             connection_config.get_saas_config().client_config.authentication  # type: ignore
         )
-        auth_strategy: OAuth2AuthenticationStrategy = get_strategy(  # type: ignore
+        auth_strategy: OAuth2AuthCodeAuthenticationStrategy = get_strategy(  # type: ignore
             authentication.strategy, authentication.configuration  # type: ignore
         )
-        auth_strategy.get_access_token(db, code, connection_config)
+        connection_config.secrets = {**connection_config.secrets, "code": code}  # type: ignore
+        auth_strategy.get_access_token(connection_config, db)
     except (OAuth2TokenException, FidesopsException) as exc:
         raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail=str(exc))
