+++ src/fidesops/ops/schemas/saas/strategy_configuration.py
@@ -113,7 +113,7 @@ class QueryParamAuthenticationConfiguration(StrategyConfiguration):
     value: str
 
 
-class OAuth2AuthenticationConfiguration(StrategyConfiguration):
+class OAuth2AuthCodeAuthenticationConfiguration(StrategyConfiguration):
     """
     OAuth2 endpoints for authentication, token retrieval, and token refresh.
     Includes an optional expires_in parameter (in seconds) for OAuth2 integrations that
@@ -122,5 +122,17 @@ class OAuth2AuthenticationConfiguration(StrategyConfiguration):
 
     expires_in: Optional[int]
     authorization_request: SaaSRequest
+    token_request: SaaSRequest
+    refresh_request: Optional[SaaSRequest]
+
+
+class OAuth2ClientCredentialsAuthenticationConfiguration(StrategyConfiguration):
+    """
+    OAuth2 endpoints for token retrieval, and token refresh.
+    Includes an optional expires_in parameter (in seconds) for OAuth2 integrations that
+    do not specify a TTL for the access tokens.
+    """
+
+    expires_in: Optional[int]
     token_request: SaaSRequest
     refresh_request: Optional[SaaSRequest]
